dt-bindings: apple,aic: Document A7-A11 compatibles
Document and describe the compatibles for Apple A7-A11 SoCs.
There are three feature levels:

 - apple,aic: No fast IPI, for A7-A10
 - apple,t8015-aic: fast IPI, global only, for A11
 - apple,t8103-aic: fast IPI with local and global support, for M1

Each feature level is an extension of the previous, for example, M1 will
also work with the A7 feature level.

All of A7-M1 gets its own SoC-specific compatible, and the "apple,aic"
compatible as a fallback.

Signed-off-by: Nick Chan <towinchenmi@gmail.com>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Reviewed-by: Sven Peter <sven@svenpeter.dev>
Acked-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>
Link: https://lore.kernel.org/all/20240901034143.12731-2-towinchenmi@gmail.com