arm64: errata: Add feature to denote Apple Fusion Architecture
The Apple A10 family consists of physical performance and efficiency
cores, and only one of them can be active at a given time depending on
the current p-state. This results in logical cores that could switch
between P-mode and E-mode. This results in the logical core being variable
depending on the current p-state.

For example, only the physical P-core can execute 32-bit EL0, and the
physical cores have different MPIDR values.

Add a feature to denote this situation, and the workarounds for this will
be added later.

Signed-off-by: Nick Chan <towinchenmi@gmail.com>