irqchip/apple-aic: Introduce CPU mapping table for AIC1
On some systems such as the Apple Development Platform (2020), the system
boots, from AIC's perspective, a CPU other than CPU0, which results in
the kernel's idea of the CPU order being out of sync with AIC's idea.

Introduce a mapping table for AIC1 to map AIC's idea of CPU order to
kernel's idea of CPU order.

Signed-off-by: Nick Chan <towinchenmi@gmail.com>