arm64: dts: apple: t8012: Add PMGR nodes
Add the two PMGR nodes and all known power state subnodes. Since there
are a large number of them, put them in a separate file to include.

On models with only 1 GB of memory, only two memory channels are used,
and on models with 2 GB of memory, four memory channels are used. The
"apple,always-on" property of the extra memory channel power domains
(ps_dcs2, ps_dcs3) will be removed by loader on models with 1 GB of
memory.

The amount of memory depends on the storage configuration of the Mac.

Signed-off-by: Nick Chan <towinchenmi@gmail.com>